third_party/WebKit/LayoutTests/http/tests/eventsource/eventsource-cors-with-credentials-expected.txt

CONSOLE ERROR: EventSource cannot load http://localhost:8000/eventsource/resources/es-cors-credentials.php?count=1. Wildcards cannot be used in the 'Access-Control-Allow-Origin' header when the credentials flag is true. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: EventSource cannot load http://localhost:8000/eventsource/resources/es-cors-credentials.php?count=2. Wildcards cannot be used in the 'Access-Control-Allow-Origin' header when the credentials flag is true. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: EventSource cannot load http://localhost:8000/eventsource/resources/es-cors-credentials.php?count=3. Credentials flag is 'true', but the 'Access-Control-Allow-Credentials' header is ''. It must be 'true' to allow credentials.
Test that EventSource cross-origin requests with credentials fail until the correct CORS headers are sent. Should print a series of PASS messages followed by DONE.

PASS: got error event and readyState is CLOSED
PASS: got error event and readyState is CLOSED
PASS: got error event and readyState is CLOSED
PASS: got cross-origin message event with data "DATA1" and lastEventId "77"
PASS: got cross-origin message event with data "DATA2"
DONE
